(statscreen) final cleanup
