+++ include/functions.h
@@ -4725,7 +4725,7 @@ void sub_8086B2C(int arg0, int arg1, u16* tilemap, int arg3, int arg4, int arg5,
 // ??? sub_808979C(???);
 void LoadDialogueBoxGfx(void* dest, int pal);
 // ??? sub_80898C4(???);
-// ??? sub_8089980(???);
+void sub_8089980(int x, int y, int w, int h, int unk);
 // ??? sub_8089C40(???);
 // ??? sub_8089CD4(???);
 // ??? sub_8089D50(???);
@@ -4736,8 +4736,8 @@ void LoadDialogueBoxGfx(void* dest, int pal);
 // ??? sub_8089F58(???);
 // ??? sub_8089FCC(???);
 // ??? sub_808A00C(???);
-// ??? sub_808A0FC(???);
-// ??? sub_808A118(???);
+void sub_808A0FC(int item, int mid);
+void sub_808A118(void);
 // ??? sub_808A160(???);
 // ??? sub_808A188(???);
 // ??? sub_808A1B8(???);
